Move functions from NewMovie down to DetailsForm
